+++ src/main/java/dev/mim1q/gimm1q/mixin/client/model/CuboidOverlayRendererMixin.java
@@ -1,8 +1,8 @@
 package dev.mim1q.gimm1q.mixin.client.model;
 
 import com.google.common.collect.ImmutableMap;
-import dev.mim1q.gimm1q.client.render.ModelOverlayVertexConsumer;
-import dev.mim1q.gimm1q.client.render.ModelOverlayVertexConsumer.OverlayUvMapperQuadContext;
+import dev.mim1q.gimm1q.client.render.overlay.ModelOverlayVertexConsumer;
+import dev.mim1q.gimm1q.client.render.overlay.OverlayUvMapper;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.render.VertexConsumer;
@@ -149,7 +149,7 @@ public class CuboidOverlayRendererMixin {
                     final var uv = modelOverlayVertexConsumer.applyUvMapper(
                         vertex.u * gimm1q$textureWidth / modelOverlayVertexConsumer.getTextureSize(),
                         vertex.v * gimm1q$textureHeight / modelOverlayVertexConsumer.getTextureSize(),
-                        new OverlayUvMapperQuadContext(
+                        new OverlayUvMapper.OverlayUvMapperQuadContext(
                             animationProgress,
                             index,
                             gimm1q$isMirrored
